@@ -64,7 +64,7 @@ class DecisionEventInfo {
   /// The configuration set containing additional information relevant to the
   /// prediction state when the current event occurred, or null if no
   /// additional information is relevant or available.
-  final ATNConfigSet configs;
+  final ATNConfigSet? configs;
 
   /// The input token stream which is being parsed.
   final TokenStream input;
@@ -80,8 +80,14 @@ class DecisionEventInfo {
   /// otherwise, [false] if the input occurred during SLL prediction.
   final bool fullCtx;
 
-  DecisionEventInfo(this.decision, this.configs, this.input, this.startIndex,
-      this.stopIndex, this.fullCtx);
+  DecisionEventInfo(
+    this.decision,
+    this.configs,
+    this.input,
+    this.startIndex,
+    this.stopIndex,
+    this.fullCtx,
+  );
 }
 
 /// This class contains profiling gathered for a particular decision.
@@ -96,12 +102,14 @@ class DecisionEventInfo {
 ///
 /// @since 4.3
 class DecisionInfo {
+
+
   /// The decision number, which is an index into {@link ATN#decisionToState}.
   final int decision;
 
   /// The total number of times {@link ParserATNSimulator#adaptivePredict} was
   /// invoked for this decision.
-  int invocations;
+  int invocations = 0;
 
   /// The total time spent in {@link ParserATNSimulator#adaptivePredict} for
   /// this decision, in nanoseconds.
@@ -115,50 +123,50 @@ class DecisionInfo {
   /// which is warmed up by parsing the input prior to profiling. If desired,
   /// call {@link ATNSimulator#clearDFA} to reset the DFA cache to its initial
   /// state before starting the profiling measurement pass.</p>
-  int timeInPrediction;
+  int timeInPrediction = 0;
 
   /// The sum of the lookahead required for SLL prediction for this decision.
   /// Note that SLL prediction is used before LL prediction for performance
   /// reasons even when {@link PredictionMode#LL} or
   /// {@link PredictionMode#LL_EXACT_AMBIG_DETECTION} is used.
-  int SLL_TotalLook;
+  int SLL_TotalLook = 0;
 
   /// Gets the minimum lookahead required for any single SLL prediction to
   /// complete for this decision, by reaching a unique prediction, reaching an
   /// SLL conflict state, or encountering a syntax error.
-  int SLL_MinLook;
+  int SLL_MinLook = 0;
 
   /// Gets the maximum lookahead required for any single SLL prediction to
   /// complete for this decision, by reaching a unique prediction, reaching an
   /// SLL conflict state, or encountering a syntax error.
-  int SLL_MaxLook;
+  int SLL_MaxLook = 0;
 
   /// Gets the [LookaheadEventInfo] associated with the event where the
   /// {@link #SLL_MaxLook} value was set.
-  LookaheadEventInfo SLL_MaxLookEvent;
+  LookaheadEventInfo? SLL_MaxLookEvent;
 
   /// The sum of the lookahead required for LL prediction for this decision.
   /// Note that LL prediction is only used when SLL prediction reaches a
   /// conflict state.
-  int LL_TotalLook;
+  int LL_TotalLook = 0;
 
   /// Gets the minimum lookahead required for any single LL prediction to
   /// complete for this decision. An LL prediction completes when the algorithm
   /// reaches a unique prediction, a conflict state (for
   /// {@link PredictionMode#LL}, an ambiguity state (for
   /// {@link PredictionMode#LL_EXACT_AMBIG_DETECTION}, or a syntax error.
-  int LL_MinLook;
+  int LL_MinLook = 0;
 
   /// Gets the maximum lookahead required for any single LL prediction to
   /// complete for this decision. An LL prediction completes when the algorithm
   /// reaches a unique prediction, a conflict state (for
   /// {@link PredictionMode#LL}, an ambiguity state (for
   /// {@link PredictionMode#LL_EXACT_AMBIG_DETECTION}, or a syntax error.
-  int LL_MaxLook;
+  int LL_MaxLook = 0;
 
   /// Gets the [LookaheadEventInfo] associated with the event where the
   /// {@link #LL_MaxLook} value was set.
-  LookaheadEventInfo LL_MaxLookEvent;
+  LookaheadEventInfo? LL_MaxLookEvent;
 
   /// A collection of [ContextSensitivityInfo] instances describing the
   /// context sensitivities encountered during LL prediction for this decision.
@@ -200,7 +208,7 @@ class DecisionInfo {
   /// @see #SLL_ATNTransitions
   /// @see ParserATNSimulator#computeTargetState
   /// @see LexerATNSimulator#computeTargetState
-  int SLL_ATNTransitions;
+  int SLL_ATNTransitions = 0;
 
   /// The total number of DFA transitions required during SLL prediction for
   /// this decision.
@@ -210,7 +218,7 @@ class DecisionInfo {
   ///
   /// @see ParserATNSimulator#getExistingTargetState
   /// @see LexerATNSimulator#getExistingTargetState
-  int SLL_DFATransitions;
+  int SLL_DFATransitions = 0;
 
   /// Gets the total number of times SLL prediction completed in a conflict
   /// state, resulting in fallback to LL prediction.
@@ -221,7 +229,7 @@ class DecisionInfo {
   /// conflicts for this decision produce the same result as LL prediction for
   /// this decision, {@link PredictionMode#SLL} would produce the same overall
   /// parsing result as {@link PredictionMode#LL}.</p>
-  int LL_Fallback;
+  int LL_Fallback = 0;
 
   /// The total number of ATN transitions required during LL prediction for
   /// this decision. An ATN transition is determined by the number of times the
@@ -237,7 +245,7 @@ class DecisionInfo {
   /// @see #LL_DFATransitions
   /// @see ParserATNSimulator#computeTargetState
   /// @see LexerATNSimulator#computeTargetState
-  int LL_ATNTransitions;
+  int LL_ATNTransitions = 0;
 
   /// The total number of DFA transitions required during LL prediction for
   /// this decision.
@@ -247,7 +255,7 @@ class DecisionInfo {
   ///
   /// @see ParserATNSimulator#getExistingTargetState
   /// @see LexerATNSimulator#getExistingTargetState
-  int LL_DFATransitions;
+  int LL_DFATransitions = 0;
 
   /// Constructs a new instance of the [DecisionInfo] class to contain
   /// statistics for a particular decision.
@@ -296,7 +304,7 @@ class DecisionInfo {
 /// @since 4.3
 class AmbiguityInfo extends DecisionEventInfo {
   /// The set of alternative numbers for this decision event that lead to a valid parse. */
-  BitSet ambigAlts;
+  BitSet? ambigAlts;
 
   /// Constructs a new instance of the [AmbiguityInfo] class with the
   /// specified detailed ambiguity information.
@@ -340,9 +348,14 @@ class ErrorInfo extends DecisionEventInfo {
   /// @param fullCtx [true] if the syntax error was identified during LL
   /// prediction; otherwise, [false] if the syntax error was identified
   /// during SLL prediction
-  ErrorInfo(int decision, ATNConfigSet configs, TokenStream input,
-      int startIndex, int stopIndex, bool fullCtx)
-      : super(decision, configs, input, startIndex, stopIndex, fullCtx);
+  ErrorInfo(
+    int decision,
+    ATNConfigSet? configs,
+    TokenStream input,
+    int startIndex,
+    int stopIndex,
+    bool fullCtx,
+  ) : super(decision, configs, input, startIndex, stopIndex, fullCtx);
 }
 
 /// This class represents profiling event information for tracking the lookahead
@@ -369,9 +382,15 @@ class LookaheadEventInfo extends DecisionEventInfo {
   /// @param fullCtx [true] if the current lookahead is part of an LL
   /// prediction; otherwise, [false] if the current lookahead is part of
   /// an SLL prediction
-  LookaheadEventInfo(int decision, ATNConfigSet configs, this.predictedAlt,
-      TokenStream input, int startIndex, int stopIndex, bool fullCtx)
-      : super(decision, configs, input, startIndex, stopIndex, fullCtx);
+  LookaheadEventInfo(
+    int decision,
+    ATNConfigSet? configs,
+    this.predictedAlt,
+    TokenStream input,
+    int startIndex,
+    int stopIndex,
+    bool fullCtx,
+  ) : super(decision, configs, input, startIndex, stopIndex, fullCtx);
 }
 
 /// This class represents profiling event information for semantic predicate
