@@ -34,7 +34,7 @@ abstract class SemanticContext {
   /// capture context dependent predicates in the context in which we begin
   /// prediction, so we passed in the outer context here in case of context
   /// dependent predicate evaluation.</p>
-  bool eval(Recognizer parser, RuleContext parserCallStack);
+  bool eval(Recognizer parser, RuleContext? parserCallStack);
 
   /// Evaluate the precedence predicates for the context and reduce the result.
   ///
@@ -52,12 +52,14 @@ abstract class SemanticContext {
   /// <li>A non-null [SemanticContext]in the new simplified
   /// semantic context after precedence predicates are evaluated.</li>
   /// </ul>
-  SemanticContext evalPrecedence(Recognizer parser,
-      RuleContext parserCallStack) {
+  SemanticContext? evalPrecedence(
+    Recognizer parser,
+    RuleContext? parserCallStack,
+  ) {
     return this;
   }
 
-  static SemanticContext and(SemanticContext a, SemanticContext b) {
+  static SemanticContext? and(SemanticContext? a, SemanticContext? b) {
     if (a == null || a == NONE) return b;
     if (b == null || b == NONE) return a;
     final result = AND(a, b);
@@ -70,7 +72,7 @@ abstract class SemanticContext {
 
   ///
   ///  @see ParserATNSimulator#getPredsForAmbigAlts
-  static SemanticContext or(SemanticContext a, SemanticContext b) {
+  static SemanticContext? or(SemanticContext? a, SemanticContext? b) {
     if (a == null) return b;
     if (b == null) return a;
     if (a == NONE || b == NONE) return NONE;
@@ -89,7 +91,7 @@ abstract class SemanticContext {
 
   static Iterable<SemanticContext> filterNonPrecedencePredicates(
       Iterable<SemanticContext> collection) {
-    return collection.where((e) => !(e is PrecedencePredicate));
+    return collection.where((e) => e is! PrecedencePredicate);
   }
 }
 
@@ -102,7 +104,7 @@ class Predicate extends SemanticContext {
       [this.ruleIndex = -1, this.predIndex = -1, this.isCtxDependent = false]);
 
   @override
-  bool eval(Recognizer parser, RuleContext parserCallStack) {
+  bool eval(Recognizer parser, RuleContext? parserCallStack) {
     final localctx = isCtxDependent ? parserCallStack : null;
     return parser.sempred(localctx, ruleIndex, predIndex);
   }
@@ -138,13 +140,15 @@ class PrecedencePredicate extends SemanticContext
   PrecedencePredicate([this.precedence = 0]);
 
   @override
-  bool eval(Recognizer parser, RuleContext parserCallStack) {
+  bool eval(Recognizer parser, RuleContext? parserCallStack) {
     return parser.precpred(parserCallStack, precedence);
   }
 
   @override
-  SemanticContext evalPrecedence(Recognizer parser,
-      RuleContext parserCallStack) {
+  SemanticContext? evalPrecedence(
+    Recognizer parser,
+    RuleContext? parserCallStack,
+  ) {
     if (parser.precpred(parserCallStack, precedence)) {
       return SemanticContext.NONE;
     } else {
@@ -165,11 +169,10 @@ class PrecedencePredicate extends SemanticContext
   }
 
   @override
-  bool operator ==(Object obj) {
-    if (!(obj is PrecedencePredicate)) {
+  bool operator ==(Object other) {
+    if (other is! PrecedencePredicate) {
       return false;
     }
-    PrecedencePredicate other = obj;
     return precedence == other.precedence;
   }
 
@@ -198,7 +201,7 @@ abstract class Operator extends SemanticContext {
 /// is false.
 
 class AND extends Operator {
-  List<SemanticContext> opnds;
+  late final List<SemanticContext> opnds;
 
   AND(SemanticContext a, SemanticContext b) {
     var operands = <SemanticContext>{};
@@ -214,13 +217,13 @@ class AND extends Operator {
     }
 
     final precedencePredicates =
-    SemanticContext.filterPrecedencePredicates(operands);
+        SemanticContext.filterPrecedencePredicates(operands);
 
     operands = SemanticContext.filterNonPrecedencePredicates(operands).toSet();
     if (precedencePredicates.isNotEmpty) {
       // interested in the transition with the lowest precedence
       final reduced =
-      precedencePredicates.reduce((a, b) => a.compareTo(b) <= 0 ? a : b);
+          precedencePredicates.reduce((a, b) => a.compareTo(b) <= 0 ? a : b);
       operands.add(reduced);
     }
 
@@ -233,9 +236,8 @@ class AND extends Operator {
   }
 
   @override
-  bool operator ==(Object obj) {
-    if (!(obj is AND)) return false;
-    AND other = obj;
+  bool operator ==(Object other) {
+    if (other is! AND) return false;
     return ListEquality().equals(opnds, other.opnds);
   }
 
@@ -251,7 +253,7 @@ class AND extends Operator {
   /// unordered.</p>
 
   @override
-  bool eval(Recognizer parser, RuleContext parserCallStack) {
+  bool eval(Recognizer parser, RuleContext? parserCallStack) {
     for (var opnd in opnds) {
       if (!opnd.eval(parser, parserCallStack)) return false;
     }
@@ -259,13 +261,14 @@ class AND extends Operator {
   }
 
   @override
-  SemanticContext evalPrecedence(Recognizer parser,
-      RuleContext parserCallStack) {
+  SemanticContext? evalPrecedence(
+    Recognizer parser,
+    RuleContext? parserCallStack,
+  ) {
     var differs = false;
     final operands = <SemanticContext>[];
     for (var context in opnds) {
-      final evaluated =
-      context.evalPrecedence(parser, parserCallStack);
+      final evaluated = context.evalPrecedence(parser, parserCallStack);
       differs |= (evaluated != context);
       if (evaluated == null) {
         // The AND context is false if any element is false
@@ -285,7 +288,7 @@ class AND extends Operator {
       return SemanticContext.NONE;
     }
 
-    var result = operands[0];
+    SemanticContext? result = operands[0];
     for (var i = 1; i < operands.length; i++) {
       result = SemanticContext.and(result, operands[i]);
     }
@@ -302,7 +305,7 @@ class AND extends Operator {
 /// A semantic context which is true whenever at least one of the contained
 /// contexts is true.
 class OR extends Operator {
-  List<SemanticContext> opnds;
+  late final List<SemanticContext> opnds;
 
   OR(SemanticContext a, SemanticContext b) {
     var operands = <SemanticContext>{};
@@ -318,13 +321,13 @@ class OR extends Operator {
     }
 
     final precedencePredicates =
-    SemanticContext.filterPrecedencePredicates(operands);
+        SemanticContext.filterPrecedencePredicates(operands);
 
     operands = SemanticContext.filterNonPrecedencePredicates(operands).toSet();
     if (precedencePredicates.isNotEmpty) {
       // interested in the transition with the highest precedence
       final reduced =
-      precedencePredicates.reduce((a, b) => a.compareTo(b) >= 0 ? a : b);
+          precedencePredicates.reduce((a, b) => a.compareTo(b) >= 0 ? a : b);
       operands.add(reduced);
     }
 
@@ -337,9 +340,8 @@ class OR extends Operator {
   }
 
   @override
-  bool operator ==(Object obj) {
-    if (!(obj is OR)) return false;
-    OR other = obj;
+  bool operator ==(Object other) {
+    if (other is! OR) return false;
     return ListEquality().equals(opnds, other.opnds);
   }
 
@@ -355,7 +357,7 @@ class OR extends Operator {
   /// unordered.</p>
 
   @override
-  bool eval(Recognizer parser, RuleContext parserCallStack) {
+  bool eval(Recognizer parser, RuleContext? parserCallStack) {
     for (var opnd in opnds) {
       if (opnd.eval(parser, parserCallStack)) return true;
     }
@@ -363,13 +365,14 @@ class OR extends Operator {
   }
 
   @override
-  SemanticContext evalPrecedence(Recognizer parser,
-      RuleContext parserCallStack) {
+  SemanticContext? evalPrecedence(
+    Recognizer parser,
+    RuleContext? parserCallStack,
+  ) {
     var differs = false;
     final operands = <SemanticContext>[];
     for (var context in opnds) {
-      final evaluated =
-      context.evalPrecedence(parser, parserCallStack);
+      final evaluated = context.evalPrecedence(parser, parserCallStack);
       differs |= (evaluated != context);
       if (evaluated == SemanticContext.NONE) {
         // The OR context is true if any element is true
@@ -389,7 +392,7 @@ class OR extends Operator {
       return null;
     }
 
-    var result = operands[0];
+    SemanticContext? result = operands[0];
     for (var i = 1; i < operands.length; i++) {
       result = SemanticContext.or(result, operands[i]);
     }
