@@ -45,7 +45,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
 
   /// The [ParserRuleContext] object for the currently executing rule.
   /// This is always non-null during the parsing process.
-  ParserRuleContext _ctx;
+  ParserRuleContext? context;
 
   /// Specifies whether or not the parser should construct a parse tree during
   /// the parsing process. The default value is [true].
@@ -59,13 +59,13 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// later call to {@link #setTrace}{@code (false)}. The listener itself is
   /// implemented as a parser listener so this field is not directly used by
   /// other parser methods.
-  TraceListener _tracer;
+  TraceListener? _tracer;
 
   /// The list of [ParseTreeListener] listeners registered to receive
   /// events during the parse.
   ///
   /// @see #addParseListener
-  List<ParseTreeListener> _parseListeners;
+  List<ParseTreeListener>? _parseListeners;
 
   /// The number of syntax errors reported during parsing. This value is
   /// incremented each time {@link #notifyErrorListeners} is called.
@@ -74,23 +74,21 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// Indicates parser has match()ed EOF token. See {@link #exitRule()}. */
   bool matchedEOF = false;
 
-  Parser(TokenStream input) {
-    inputStream = input;
+  Parser(this._input) {
+    reset(false);
   }
 
   /// reset the parser's state */
-  void reset() {
-    if (inputStream != null) inputStream.seek(0);
+  void reset([bool resetInput = true]) {
+    if (resetInput) inputStream.seek(0);
     errorHandler.reset(this);
-    _ctx = null;
+    context = null;
     _syntaxErrors = 0;
     matchedEOF = false;
     setTrace(false);
     _precedenceStack.clear();
     _precedenceStack.add(0);
-    if (interpreter != null) {
-      interpreter.reset();
-    }
+    interpreter?.reset();
   }
 
   /// Match current input symbol against [ttype]. If the symbol type
@@ -123,7 +121,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
       if (_buildParseTrees && t.tokenIndex == -1) {
         // we must have conjured up a new token during single token insertion
         // if it's not the current symbol
-        _ctx.addErrorNode(createErrorNode(_ctx, t));
+        context!.addErrorNode(createErrorNode(context!, t));
       }
     }
     return t;
@@ -155,7 +153,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
       if (_buildParseTrees && t.tokenIndex == -1) {
         // we must have conjured up a new token during single token insertion
         // if it's not the current symbol
-        _ctx.addErrorNode(createErrorNode(_ctx, t));
+        context!.addErrorNode(createErrorNode(context!, t));
       }
     }
 
@@ -205,10 +203,10 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// @return [true] if the {@link ParserRuleContext#children} list is trimmed
   /// using the default {@link Parser.TrimToSizeListener} during the parse process.
   bool get trimParseTree {
-    return parseListeners.contains(TrimToSizeListener.INSTANCE);
+    return parseListeners?.contains(TrimToSizeListener.INSTANCE) ?? false;
   }
 
-  List<ParseTreeListener> get parseListeners => _parseListeners;
+  List<ParseTreeListener>? get parseListeners => _parseListeners;
 
   /// Registers [listener] to receive events during the parsing process.
   ///
@@ -237,14 +235,12 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// @param listener the listener to add
   ///
   /// @throws NullPointerException if {@code} listener is null
-  void addParseListener(ParseTreeListener listener) {
-    if (listener == null) {
-      throw ArgumentError.notNull('listener');
-    }
-
+  void addParseListener(
+    ParseTreeListener listener,
+  ) {
     _parseListeners ??= [];
 
-    _parseListeners.add(listener);
+    _parseListeners!.add(listener);
   }
 
   /// Remove [listener] from the list of parse listeners.
@@ -255,10 +251,10 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// @see #addParseListener
   ///
   /// @param listener the listener to remove
-  void removeParseListener(ParseTreeListener listener) {
+  void removeParseListener(ParseTreeListener? listener) {
     if (_parseListeners != null) {
-      if (_parseListeners.remove(listener)) {
-        if (_parseListeners.isEmpty) {
+      if (_parseListeners!.remove(listener)) {
+        if (_parseListeners!.isEmpty) {
           _parseListeners = null;
         }
       }
@@ -276,9 +272,10 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   ///
   /// @see #addParseListener
   void triggerEnterRuleEvent() {
-    for (var listener in _parseListeners) {
-      listener.enterEveryRule(_ctx);
-      _ctx.enterRule(listener);
+    if (_parseListeners == null) return;
+    for (var listener in _parseListeners!) {
+      listener.enterEveryRule(context!);
+      context!.enterRule(listener);
     }
   }
 
@@ -286,11 +283,12 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   ///
   /// @see #addParseListener
   void triggerExitRuleEvent() {
+    if (_parseListeners == null) return;
     // reverse order walk of listeners
-    for (var i = _parseListeners.length - 1; i >= 0; i--) {
-      final listener = _parseListeners[i];
-      _ctx.exitRule(listener);
-      listener.exitEveryRule(_ctx);
+    for (var i = _parseListeners!.length - 1; i >= 0; i--) {
+      final listener = _parseListeners![i];
+      context!.exitRule(listener);
+      listener.exitEveryRule(context!);
     }
   }
 
@@ -319,20 +317,18 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// @throws UnsupportedOperationException if the current parser does not
   /// implement the {@link #getSerializedATN()} method.
   ATN get ATNWithBypassAlts {
-    final serializedAtn = serializedATN;
-    if (serializedAtn == null) {
+    if (serializedATN == null) {
       throw UnsupportedError(
           'The current parser does not support an ATN with bypass alternatives.');
     }
 
-    var result = bypassAltsAtnCache[serializedAtn];
+    var result = bypassAltsAtnCache[serializedATN];
     if (result == null) {
-      final deserializationOptions =
-          ATNDeserializationOptions();
+      final deserializationOptions = ATNDeserializationOptions(false);
       deserializationOptions.setGenerateRuleBypassTransitions(true);
       result = ATNDeserializer(deserializationOptions)
-          .deserialize(serializedAtn.codeUnits);
-      bypassAltsAtnCache[serializedAtn] = result;
+          .deserialize(serializedATN!.codeUnits);
+      bypassAltsAtnCache[serializedATN!] = result;
     }
 
     return result;
@@ -347,11 +343,14 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// ParseTreeMatch m = p.match(t);
   /// String id = m.get("ID");
   /// </pre>
-  ParseTreePattern compileParseTreePattern(String pattern, int patternRuleIndex,
-      [Lexer lexer]) {
+  ParseTreePattern compileParseTreePattern(
+    String pattern,
+    int patternRuleIndex, [
+    Lexer? lexer,
+  ]) {
     if (lexer == null) {
-      final tokenSource = tokenStream?.tokenSource;
-      if (tokenSource == null || !(tokenSource is Lexer)) {
+      final tokenSource = tokenStream.tokenSource;
+      if (tokenSource is! Lexer) {
         throw UnsupportedError("Parser can't discover a lexer to use");
       }
       lexer = tokenSource;
@@ -365,7 +364,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   TokenStream get inputStream => tokenStream;
 
   @override
-  set inputStream(IntStream input) {
+  set inputStream(TokenStream input) {
     setTokenStream(input);
   }
 
@@ -373,8 +372,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
 
   /// Set the token stream and reset the parser. */
   void setTokenStream(TokenStream input) {
-    _input = null;
-    reset();
+    reset(false);
     _input = input;
   }
 
@@ -382,21 +380,30 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   ///  into the label for the associated token ref; e.g., x=ID.
 
   Token get currentToken {
-    return _input.LT(1);
+    return _input.LT(1)!;
   }
 
-  void notifyErrorListeners(String msg,
-      [Token offendingToken, RecognitionException e]) {
+  void notifyErrorListeners(
+    String msg, [
+    Token? offendingToken,
+    RecognitionException? e,
+  ]) {
     offendingToken = offendingToken ?? currentToken;
     _syntaxErrors++;
-    var line = -1;
+    int? line = -1;
     var charPositionInLine = -1;
     line = offendingToken.line;
     charPositionInLine = offendingToken.charPositionInLine;
 
     final listener = errorListenerDispatch;
     listener.syntaxError(
-        this, offendingToken, line, charPositionInLine, msg, e);
+      this,
+      offendingToken,
+      line,
+      charPositionInLine,
+      msg,
+      e,
+    );
   }
 
   /// Consume and return the {@linkplain #getCurrentToken current symbol}.
@@ -423,19 +430,19 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
     if (o.type != IntStream.EOF) {
       inputStream.consume();
     }
-    final hasListener = _parseListeners != null && _parseListeners.isNotEmpty;
+    final hasListener = _parseListeners != null && _parseListeners!.isNotEmpty;
     if (_buildParseTrees || hasListener) {
       if (errorHandler.inErrorRecoveryMode(this)) {
-        final node = _ctx.addErrorNode(createErrorNode(_ctx, o));
+        final node = context!.addErrorNode(createErrorNode(context!, o));
         if (_parseListeners != null) {
-          for (var listener in _parseListeners) {
+          for (var listener in _parseListeners!) {
             listener.visitErrorNode(node);
           }
         }
       } else {
-        final node = _ctx.addChild(createTerminalNode(_ctx, o));
+        final node = context!.addChild(createTerminalNode(context!, o));
         if (_parseListeners != null) {
-          for (var listener in _parseListeners) {
+          for (var listener in _parseListeners!) {
             listener.visitTerminal(node);
           }
         }
@@ -461,10 +468,10 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   }
 
   void addContextToParseTree() {
-    final parent = _ctx.parent;
+    final parent = context?.parent;
     // add current context to parent if we have a parent
     if (parent != null) {
-      parent.addAnyChild(_ctx);
+      parent.addAnyChild(context!);
     }
   }
 
@@ -472,37 +479,39 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// {@link #_ctx} get the current context.
   void enterRule(ParserRuleContext localctx, int state, int ruleIndex) {
     this.state = state;
-    _ctx = localctx;
-    _ctx.start = _input.LT(1);
+    context = localctx;
+    context!.start = _input.LT(1)!;
     if (_buildParseTrees) addContextToParseTree();
     if (_parseListeners != null) triggerEnterRuleEvent();
   }
 
   void exitRule() {
+    assert(context != null);
     if (matchedEOF) {
       // if we have matched EOF, it cannot consume past EOF so we use LT(1) here
-      _ctx.stop = _input.LT(1); // LT(1) will be end of file
+      context!.stop = _input.LT(1); // LT(1) will be end of file
     } else {
-      _ctx.stop = _input.LT(-1); // stop node is what we just matched
+      context!.stop = _input.LT(-1); // stop node is what we just matched
     }
     // trigger event on _ctx, before it reverts to parent
     if (_parseListeners != null) triggerExitRuleEvent();
-    state = _ctx.invokingState;
-    _ctx = _ctx.parent;
+    state = context!.invokingState;
+    context = context?.parent;
   }
 
   void enterOuterAlt(ParserRuleContext localctx, int altNum) {
+    assert(context != null);
     localctx.altNumber = altNum;
     // if we have new localctx, make sure we replace existing ctx
     // that is previous child of parse tree
-    if (_buildParseTrees && _ctx != localctx) {
-      final parent = _ctx.parent;
+    if (_buildParseTrees && context != localctx) {
+      final parent = context!.parent;
       if (parent != null) {
         parent.removeLastChild();
         parent.addAnyChild(localctx);
       }
     }
-    _ctx = localctx;
+    context = localctx;
   }
 
   /// Get the precedence level for the top-most precedence rule.
@@ -521,8 +530,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
       ParserRuleContext localctx, int state, int ruleIndex, int precedence) {
     this.state = state;
     _precedenceStack.add(precedence);
-    _ctx = localctx;
-    _ctx.start = _input.LT(1);
+    context = localctx;
+    context!.start = _input.LT(1)!;
     if (_parseListeners != null) {
       triggerEnterRuleEvent(); // simulates rule entry for left-recursive rules
     }
@@ -531,16 +540,20 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// Like {@link #enterRule} but for recursive rules.
   ///  Make the current context the child of the incoming localctx.
   void pushNewRecursionContext(
-      ParserRuleContext localctx, int state, int ruleIndex) {
-    final previous = _ctx;
+    ParserRuleContext localctx,
+    int state,
+    int? ruleIndex,
+  ) {
+    assert(context != null);
+    final previous = context!;
     previous.parent = localctx;
     previous.invokingState = state;
     previous.stop = _input.LT(-1);
 
-    _ctx = localctx;
-    _ctx.start = previous.start;
+    context = localctx;
+    context!.start = previous.start;
     if (_buildParseTrees) {
-      _ctx.addAnyChild(previous);
+      context!.addAnyChild(previous);
     }
 
     if (_parseListeners != null) {
@@ -548,19 +561,20 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
     }
   }
 
-  void unrollRecursionContexts(ParserRuleContext _parentctx) {
+  void unrollRecursionContexts(ParserRuleContext? _parentctx) {
+    assert(context != null);
     _precedenceStack.removeLast();
-    _ctx.stop = _input.LT(-1);
-    final retctx = _ctx; // save current ctx (return value)
+    context!.stop = _input.LT(-1);
+    final retctx = context!; // save current ctx (return value)
 
     // unroll so _ctx is as it was before call to recursive method
     if (_parseListeners != null) {
-      while (_ctx != _parentctx) {
+      while (context != _parentctx) {
         triggerExitRuleEvent();
-        _ctx = _ctx.parent;
+        context = context!.parent;
       }
     } else {
-      _ctx = _parentctx;
+      context = _parentctx;
     }
 
     // hook into tree
@@ -572,8 +586,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
     }
   }
 
-  ParserRuleContext getInvokingContext(int ruleIndex) {
-    var p = _ctx;
+  ParserRuleContext? getInvokingContext(int ruleIndex) {
+    var p = context;
     while (p != null) {
       if (p.ruleIndex == ruleIndex) return p;
       p = p.parent;
@@ -581,16 +595,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
     return null;
   }
 
-  ParserRuleContext get context {
-    return _ctx;
-  }
-
-  set context(ParserRuleContext ctx) {
-    _ctx = ctx;
-  }
-
   @override
-  bool precpred(RuleContext localctx, int precedence) {
+  bool precpred(RuleContext? localctx, int precedence) {
     return precedence >= _precedenceStack.last;
   }
 
@@ -612,11 +618,11 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// @return [true] if [symbol] can follow the current state in
   /// the ATN, otherwise [false].
   bool isExpectedToken(int symbol) {
-//   		return interpreter.atn.nextTokens(_ctx);
-    final atn = interpreter.atn;
-    var ctx = _ctx;
+//   		return interpreter!.atn.nextTokens(_ctx);
+    final atn = interpreter!.atn;
+    var ctx = context;
     final s = atn.states[state];
-    var following = atn.nextTokens(s);
+    var following = atn.nextTokens(s!);
     if (following.contains(symbol)) {
       return true;
     }
@@ -626,8 +632,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
     while (ctx != null &&
         ctx.invokingState >= 0 &&
         following.contains(Token.EPSILON)) {
-      final invokingState = atn.states[ctx.invokingState];
-      RuleTransition rt = invokingState.transition(0);
+      final invokingState = atn.states[ctx.invokingState]!;
+      final rt = invokingState.transition(0) as RuleTransition;
       following = atn.nextTokens(rt.followState);
       if (following.contains(symbol)) {
         return true;
@@ -657,8 +663,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   }
 
   IntervalSet get expectedTokensWithinCurrentRule {
-    final atn = interpreter.atn;
-    final s = atn.states[state];
+    final atn = interpreter!.atn;
+    final s = atn.states[state]!;
     return atn.nextTokens(s);
   }
 
@@ -670,7 +676,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   }
 
   ParserRuleContext get ruleContext {
-    return _ctx;
+    assert(context != null);
+    return context!;
   }
 
   List<String> get ruleInvocationStack => getRuleInvocationStack();
@@ -681,8 +688,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   ///  in the ATN a rule is invoked.
   ///
   ///  This is very useful for error messages.
-  List<String> getRuleInvocationStack([RuleContext p]) {
-    p = p ?? _ctx;
+  List<String> getRuleInvocationStack([RuleContext? p]) {
+    p = p ?? context;
     final _ruleNames = ruleNames;
     final stack = <String>[];
     while (p != null) {
@@ -701,8 +708,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// For debugging and other purposes. */
   List<String> get dfaStrings {
     final s = <String>[];
-    for (var d = 0; d < interpreter.decisionToDFA.length; d++) {
-      final dfa = interpreter.decisionToDFA[d];
+    for (var d = 0; d < interpreter!.decisionToDFA.length; d++) {
+      final dfa = interpreter!.decisionToDFA[d];
       s.add(dfa.toString(vocabulary));
     }
     return s;
@@ -711,8 +718,8 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   /// For debugging and other purposes. */
   void dumpDFA() {
     var seenOne = false;
-    for (var d = 0; d < interpreter.decisionToDFA.length; d++) {
-      final dfa = interpreter.decisionToDFA[d];
+    for (var d = 0; d < interpreter!.decisionToDFA.length; d++) {
+      final dfa = interpreter!.decisionToDFA[d];
       if (dfa.states.isNotEmpty) {
         if (seenOne) print('');
         print('Decision ${dfa.decision}:');
@@ -727,7 +734,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
   }
 
   @override
-  ParseInfo get parseInfo {
+  ParseInfo? get parseInfo {
     final interp = interpreter;
     if (interp is ProfilingATNSimulator) {
       return ParseInfo(interp);
@@ -737,7 +744,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
 
   /// @since 4.3
   void setProfile(bool profile) {
-    final interp = interpreter;
+    final interp = interpreter!;
     final saveMode = interp.predictionMode;
     if (profile) {
       if (!(interp is ProfilingATNSimulator)) {
@@ -745,10 +752,14 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
       }
     } else if (interp is ProfilingATNSimulator) {
       final sim = ParserATNSimulator(
-          this, getATN(), interp.decisionToDFA, interp.sharedContextCache);
+        this,
+        getATN(),
+        interp.decisionToDFA,
+        interp.sharedContextCache,
+      );
       interpreter = sim;
     }
-    interpreter.predictionMode = saveMode;
+    interpreter!.predictionMode = saveMode;
   }
 
   /// During a parse is sometimes useful to listen in on the rule entry and exit
@@ -763,7 +774,7 @@ abstract class Parser extends Recognizer<ParserATNSimulator> {
       } else {
         _tracer = TraceListener(this);
       }
-      addParseListener(_tracer);
+      addParseListener(_tracer!);
     }
   }
 
