@@ -33,9 +33,8 @@ class ParserInterpreter extends Parser {
   final String grammarFileName;
   final ATN atn;
 
-  List<DFA> decisionToDFA; // not shared like it is for generated parsers
-  final PredictionContextCache sharedContextCache =
-      PredictionContextCache();
+  late List<DFA> decisionToDFA; // not shared like it is for generated parsers
+  final PredictionContextCache sharedContextCache = PredictionContextCache();
 
   @override
   final List<String> ruleNames;
@@ -55,7 +54,7 @@ class ParserInterpreter extends Parser {
   ///
   ///  Those values are used to create new recursive rule invocation contexts
   ///  associated with left operand of an alt like "expr '*' expr".
-  final DoubleLinkedQueue<Pair<ParserRuleContext, int>> _parentContextStack =
+  final DoubleLinkedQueue<Pair<ParserRuleContext?, int>> _parentContextStack =
       DoubleLinkedQueue();
 
   /// We need a map from (decision,inputIndex)->forced alt for computing ambiguous
@@ -69,7 +68,7 @@ class ParserInterpreter extends Parser {
   /// What is the current context when we override a decisions?  This tells
   ///  us what the root of the parse tree is when using override
   ///  for an ambiguity/lookahead check.
-  InterpreterRuleContext overrideDecisionRoot;
+  InterpreterRuleContext? overrideDecisionRoot;
 
   /// Return the root of the parse, which can be useful if the parser
   ///  bails out. You still can access the top node. Note that,
@@ -78,27 +77,34 @@ class ParserInterpreter extends Parser {
   ///  called and left recursive rule that fails.
   ///
   /// @since 4.5.1
-  InterpreterRuleContext rootContext;
+  late InterpreterRuleContext rootContext;
 
-  ParserInterpreter(this.grammarFileName, this.vocabulary, this.ruleNames,
-      this.atn, TokenStream input)
-      : super(input) {
+  ParserInterpreter(
+    this.grammarFileName,
+    this.vocabulary,
+    this.ruleNames,
+    this.atn,
+    TokenStream input,
+  ) : super(input) {
     // init decision DFA
     final numberOfDecisions = atn.numberOfDecisions;
-    decisionToDFA = List<DFA>(numberOfDecisions);
-    for (var i = 0; i < numberOfDecisions; i++) {
-      final decisionState = atn.getDecisionState(i);
-      decisionToDFA[i] = DFA(decisionState, i);
-    }
+    decisionToDFA = List<DFA>.generate(numberOfDecisions, (n) {
+      final decisionState = atn.getDecisionState(n);
+      return DFA(decisionState, n);
+    });
 
     // get atn simulator that knows how to do predictions
-    interpreter =
-        ParserATNSimulator(this, atn, decisionToDFA, sharedContextCache);
+    interpreter = ParserATNSimulator(
+      this,
+      atn,
+      decisionToDFA,
+      sharedContextCache,
+    );
   }
 
   @override
-  void reset() {
-    super.reset();
+  void reset([bool resetInput = true]) {
+    super.reset(resetInput);
     overrideDecisionReached = false;
     overrideDecisionRoot = null;
   }
@@ -113,7 +119,10 @@ class ParserInterpreter extends Parser {
     final startRuleStartState = atn.ruleToStartState[startRuleIndex];
 
     rootContext = createInterpreterRuleContext(
-        null, ATNState.INVALID_STATE_NUMBER, startRuleIndex);
+      null,
+      ATNState.INVALID_STATE_NUMBER,
+      startRuleIndex,
+    );
     if (startRuleStartState.isLeftRecursiveRule) {
       enterRecursionRule(
           rootContext, startRuleStartState.stateNumber, startRuleIndex, 0);
@@ -126,11 +135,10 @@ class ParserInterpreter extends Parser {
       switch (p.stateType) {
         case StateType.RULE_STOP:
           // pop; return from rule
-          if (context.isEmpty) {
+          if (context!.isEmpty) {
             if (startRuleStartState.isLeftRecursiveRule) {
-              final result = context;
-              final parentContext =
-                  _parentContextStack.removeLast();
+              final result = context!;
+              final parentContext = _parentContextStack.removeLast();
               unrollRecursionContexts(parentContext.a);
               return result;
             } else {
@@ -147,7 +155,7 @@ class ParserInterpreter extends Parser {
             visitState(p);
           } on RecognitionException catch (e) {
             state = atn.ruleToStopState[p.ruleIndex].stateNumber;
-            context.exception = e;
+            context!.exception = e;
             errorHandler.reportError(this, e);
             recover(e);
           }
@@ -159,18 +167,22 @@ class ParserInterpreter extends Parser {
 
   @override
   void enterRecursionRule(
-      ParserRuleContext localctx, int state, int ruleIndex, int precedence) {
-    final pair =
-        Pair<ParserRuleContext, int>(context, localctx.invokingState);
+    ParserRuleContext localctx,
+    int state,
+    int ruleIndex,
+    int precedence,
+  ) {
+    final pair = Pair<ParserRuleContext?, int>(context, localctx.invokingState);
     _parentContextStack.add(pair);
     super.enterRecursionRule(localctx, state, ruleIndex, precedence);
   }
 
   ATNState get atnState {
-    return atn.states[state];
+    return atn.states[state]!;
   }
 
   void visitState(ATNState p) {
+    assert(context != null);
 //		System.out.println("visitState "+p.stateNumber);
     var predictedAlt = 1;
     if (p is DecisionState) {
@@ -182,15 +194,19 @@ class ParserInterpreter extends Parser {
       case TransitionType.EPSILON:
         if (p.stateType == StateType.STAR_LOOP_ENTRY &&
             (p as StarLoopEntryState).isPrecedenceDecision &&
-            !(transition.target is LoopEndState)) {
+            (transition.target is! LoopEndState)) {
           // We are at the start of a left recursive rule's (...)* loop
           // and we're not taking the exit branch of loop.
           final localctx = createInterpreterRuleContext(
-              _parentContextStack.last.a,
-              _parentContextStack.last.b,
-              context.ruleIndex);
-          pushNewRecursionContext(localctx,
-              atn.ruleToStartState[p.ruleIndex].stateNumber, context.ruleIndex);
+            _parentContextStack.last.a,
+            _parentContextStack.last.b,
+            context!.ruleIndex,
+          );
+          pushNewRecursionContext(
+            localctx,
+            atn.ruleToStartState[p.ruleIndex].stateNumber,
+            context!.ruleIndex,
+          );
         }
         break;
 
@@ -202,7 +218,10 @@ class ParserInterpreter extends Parser {
       case TransitionType.SET:
       case TransitionType.NOT_SET:
         if (!transition.matches(
-            inputStream.LA(1), Token.MIN_USER_TOKEN_TYPE, 65535)) {
+          inputStream.LA(1)!,
+          Token.MIN_USER_TOKEN_TYPE,
+          65535,
+        )) {
           recoverInline();
         }
         matchWildcard();
@@ -213,7 +232,7 @@ class ParserInterpreter extends Parser {
         break;
 
       case TransitionType.RULE:
-        RuleStartState ruleStartState = transition.target;
+        final ruleStartState = transition.target as RuleStartState;
         final ruleIndex = ruleStartState.ruleIndex;
         final newctx =
             createInterpreterRuleContext(context, p.stateNumber, ruleIndex);
@@ -226,7 +245,7 @@ class ParserInterpreter extends Parser {
         break;
 
       case TransitionType.PREDICATE:
-        PredicateTransition predicateTransition = transition;
+        final predicateTransition = transition as PredicateTransition;
         if (!sempred(context, predicateTransition.ruleIndex,
             predicateTransition.predIndex)) {
           throw FailedPredicateException(this);
@@ -235,16 +254,21 @@ class ParserInterpreter extends Parser {
         break;
 
       case TransitionType.ACTION:
-        ActionTransition actionTransition = transition;
+        final actionTransition = transition as ActionTransition;
         action(
-            context, actionTransition.ruleIndex, actionTransition.actionIndex);
+          context,
+          actionTransition.ruleIndex,
+          actionTransition.actionIndex,
+        );
         break;
 
       case TransitionType.PRECEDENCE:
-        if (!precpred(context,
-            (transition as PrecedencePredicateTransition).precedence)) {
-          throw FailedPredicateException(this,
-              'precpred(context, ${(transition as PrecedencePredicateTransition).precedence})');
+        if (!precpred(
+          context,
+          (transition as PrecedencePredicateTransition).precedence,
+        )) {
+          throw FailedPredicateException(
+              this, 'precpred(context, ${(transition).precedence})');
         }
         break;
 
@@ -260,6 +284,7 @@ class ParserInterpreter extends Parser {
   ///  for subclasses to track interesting things.
   int visitDecisionState(DecisionState p) {
     var predictedAlt = 1;
+    assert(context != null);
     if (p.numberOfTransitions > 1) {
       errorHandler.sync(this);
       final decision = p.decision;
@@ -269,8 +294,11 @@ class ParserInterpreter extends Parser {
         predictedAlt = overrideDecisionAlt;
         overrideDecisionReached = true;
       } else {
-        predictedAlt =
-            interpreter.adaptivePredict(inputStream, decision, context);
+        predictedAlt = interpreter!.adaptivePredict(
+          inputStream,
+          decision,
+          context!,
+        );
       }
     }
     return predictedAlt;
@@ -279,22 +307,24 @@ class ParserInterpreter extends Parser {
   /// Provide simple "factory" for InterpreterRuleContext's.
   ///  @since 4.5.1
   InterpreterRuleContext createInterpreterRuleContext(
-      ParserRuleContext parent, int invokingStateNumber, int ruleIndex) {
+    ParserRuleContext? parent,
+    int invokingStateNumber,
+    int ruleIndex,
+  ) {
     return InterpreterRuleContext(parent, invokingStateNumber, ruleIndex);
   }
 
   void visitRuleStopState(ATNState p) {
     final ruleStartState = atn.ruleToStartState[p.ruleIndex];
     if (ruleStartState.isLeftRecursiveRule) {
-      final parentContext =
-          _parentContextStack.removeLast();
+      final parentContext = _parentContextStack.removeLast();
       unrollRecursionContexts(parentContext.a);
       state = parentContext.b;
     } else {
       exitRule();
     }
 
-    RuleTransition ruleTransition = atn.states[state].transition(0);
+    final ruleTransition = atn.states[state]!.transition(0) as RuleTransition;
     state = ruleTransition.followState.stateNumber;
   }
 
@@ -349,39 +379,43 @@ class ParserInterpreter extends Parser {
   void recover(RecognitionException e) {
     final i = inputStream.index;
     errorHandler.recover(this, e);
+    assert(this.context != null);
+    final context = this.context as ParserRuleContext;
     if (inputStream.index == i) {
       // no input consumed, better add an error node
       if (e is InputMismatchException) {
         final ime = e;
         final tok = e.offendingToken;
         var expectedTokenType = Token.INVALID_TYPE;
-        if (!ime.expectedTokens.isNil) {
-          expectedTokenType = ime.expectedTokens.minElement; // get any element
+        if (ime.expectedTokens != null && !ime.expectedTokens!.isNil) {
+          expectedTokenType = ime.expectedTokens!.minElement; // get any element
         }
         final errToken = tokenFactory.create(
-            expectedTokenType,
-            tok.text,
-            Pair(tok.tokenSource, tok.tokenSource.inputStream),
-            Token.DEFAULT_CHANNEL,
-            -1,
-            -1,
-            // invalid start/stop
-            tok.line,
-            tok.charPositionInLine);
+          expectedTokenType,
+          tok.text,
+          Pair(tok.tokenSource, tok.tokenSource?.inputStream),
+          Token.DEFAULT_CHANNEL,
+          -1,
+          -1,
+          // invalid start/stop
+          tok.line,
+          tok.charPositionInLine,
+        );
         context.addErrorNode(createErrorNode(context, errToken));
       } else {
         // NoViableAlt
         final tok = e.offendingToken;
         final errToken = tokenFactory.create(
-            Token.INVALID_TYPE,
-            tok.text,
-            Pair(tok.tokenSource, tok.tokenSource.inputStream),
-            Token.DEFAULT_CHANNEL,
-            -1,
-            -1,
-            // invalid start/stop
-            tok.line,
-            tok.charPositionInLine);
+          Token.INVALID_TYPE,
+          tok.text,
+          Pair(tok.tokenSource, tok.tokenSource?.inputStream),
+          Token.DEFAULT_CHANNEL,
+          -1,
+          -1,
+          // invalid start/stop
+          tok.line,
+          tok.charPositionInLine,
+        );
         context.addErrorNode(createErrorNode(context, errToken));
       }
     }
