@@ -33,7 +33,7 @@ class TagChunk extends Chunk {
 
   /// The label assigned to this chunk, or null if no label is
   /// assigned to the chunk.
-  final String label;
+  final String? label;
 
   /// Construct a new instance of [TagChunk] using the specified label
   /// and tag.
@@ -45,8 +45,8 @@ class TagChunk extends Chunk {
   ///
   /// @exception ArgumentError if [tag] is null or empty.
   TagChunk(this.tag, {this.label}) {
-    if (tag == null || tag.isEmpty) {
-      throw ArgumentError.value(tag, 'tag', 'cannot be null or empty');
+    if (tag.isEmpty) {
+      throw ArgumentError.value(tag, 'tag', 'cannot be empty');
     }
   }
 
@@ -56,7 +56,7 @@ class TagChunk extends Chunk {
   @override
   String toString() {
     if (label != null) {
-      return label + ':' + tag;
+      return label! + ':' + tag;
     }
 
     return tag;
@@ -73,11 +73,7 @@ class TextChunk extends Chunk {
   ///
   /// @param text The text of this chunk.
   /// @exception IllegalArgumentException if [text] is null.
-  TextChunk(this.text) {
-    if (text == null) {
-      throw ArgumentError.notNull('text');
-    }
-  }
+  TextChunk(this.text);
 
   /// {@inheritDoc}
   ///
