@@ -45,8 +45,14 @@ abstract class ErrorListener {
   ///        the reporting of an error. It is null in the case where
   ///        the parser was able to recover in line without exiting the
   ///        surrounding rule.
-  void syntaxError(Recognizer recognizer, Object offendingSymbol, int line,
-      int charPositionInLine, String msg, RecognitionException e);
+  void syntaxError(
+    Recognizer recognizer,
+    Object? offendingSymbol,
+    int? line,
+    int charPositionInLine,
+    String msg,
+    RecognitionException? e,
+  );
 
   /// This method is called by the parser when a full-context prediction
   /// results in an ambiguity.
@@ -86,8 +92,15 @@ abstract class ErrorListener {
   /// set of represented alternatives in [configs]
   /// @param configs the ATN configuration set where the ambiguity was
   /// identified
-  void reportAmbiguity(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, bool exact, BitSet ambigAlts, ATNConfigSet configs);
+  void reportAmbiguity(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    bool exact,
+    BitSet? ambigAlts,
+    ATNConfigSet configs,
+  );
 
   /// This method is called when an SLL conflict occurs and the parser is about
   /// to use the full context information to make an LL decision.
@@ -110,8 +123,14 @@ abstract class ErrorListener {
   /// as null).
   /// @param configs the ATN configuration set where the SLL conflict was
   /// detected
-  void reportAttemptingFullContext(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, BitSet conflictingAlts, ATNConfigSet configs);
+  void reportAttemptingFullContext(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    BitSet? conflictingAlts,
+    ATNConfigSet configs,
+  );
 
   /// This method is called by the parser when a full-context prediction has a
   /// unique result.
@@ -149,26 +168,57 @@ abstract class ErrorListener {
   /// @param prediction the unambiguous result of the full-context prediction
   /// @param configs the ATN configuration set where the unambiguous prediction
   /// was determined
-  void reportContextSensitivity(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, int prediction, ATNConfigSet configs);
+  void reportContextSensitivity(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    int prediction,
+    ATNConfigSet configs,
+  );
 }
 
 class BaseErrorListener extends ErrorListener {
   @override
-  void reportAmbiguity(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, bool exact, BitSet ambigAlts, ATNConfigSet configs) {}
+  void reportAmbiguity(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    bool exact,
+    BitSet? ambigAlts,
+    ATNConfigSet configs,
+  ) {}
 
   @override
-  void reportAttemptingFullContext(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, BitSet conflictingAlts, ATNConfigSet configs) {}
+  void reportAttemptingFullContext(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    BitSet? conflictingAlts,
+    ATNConfigSet configs,
+  ) {}
 
   @override
-  void reportContextSensitivity(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, int prediction, ATNConfigSet configs) {}
+  void reportContextSensitivity(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    int prediction,
+    ATNConfigSet configs,
+  ) {}
 
   @override
-  void syntaxError(Recognizer<ATNSimulator> recognizer, Object offendingSymbol,
-      int line, int charPositionInLine, String msg, RecognitionException e) {}
+  void syntaxError(
+    Recognizer<ATNSimulator> recognizer,
+    Object? offendingSymbol,
+    int? line,
+    int charPositionInLine,
+    String msg,
+    RecognitionException? e,
+  ) {}
 }
 
 class ConsoleErrorListener extends BaseErrorListener {
@@ -198,23 +248,43 @@ class ProxyErrorListener implements ErrorListener {
   final List<ErrorListener> delegates;
 
   ProxyErrorListener(this.delegates) {
+    // ignore: unnecessary_null_comparison
     if (delegates == null) {
       throw ArgumentError.notNull('delegates');
     }
   }
 
   @override
-  void syntaxError(Recognizer recognizer, Object offendingSymbol, int line,
-      int charPositionInLine, String msg, RecognitionException e) {
+  void syntaxError(
+    Recognizer recognizer,
+    Object? offendingSymbol,
+    int? line,
+    int charPositionInLine,
+    String msg,
+    RecognitionException? e,
+  ) {
     for (final listener in delegates) {
       listener.syntaxError(
-          recognizer, offendingSymbol, line, charPositionInLine, msg, e);
+        recognizer,
+        offendingSymbol,
+        line,
+        charPositionInLine,
+        msg,
+        e,
+      );
     }
   }
 
   @override
-  void reportAmbiguity(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, bool exact, BitSet ambigAlts, ATNConfigSet configs) {
+  void reportAmbiguity(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    bool exact,
+    BitSet? ambigAlts,
+    ATNConfigSet configs,
+  ) {
     for (final listener in delegates) {
       listener.reportAmbiguity(
           recognizer, dfa, startIndex, stopIndex, exact, ambigAlts, configs);
@@ -222,20 +292,44 @@ class ProxyErrorListener implements ErrorListener {
   }
 
   @override
-  void reportAttemptingFullContext(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, BitSet conflictingAlts, ATNConfigSet configs) {
+  void reportAttemptingFullContext(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    BitSet? conflictingAlts,
+    ATNConfigSet configs,
+  ) {
     for (final listener in delegates) {
       listener.reportAttemptingFullContext(
-          recognizer, dfa, startIndex, stopIndex, conflictingAlts, configs);
+        recognizer,
+        dfa,
+        startIndex,
+        stopIndex,
+        conflictingAlts,
+        configs,
+      );
     }
   }
 
   @override
-  void reportContextSensitivity(Parser recognizer, DFA dfa, int startIndex,
-      int stopIndex, int prediction, ATNConfigSet configs) {
+  void reportContextSensitivity(
+    Parser recognizer,
+    DFA dfa,
+    int startIndex,
+    int stopIndex,
+    int prediction,
+    ATNConfigSet configs,
+  ) {
     for (final listener in delegates) {
       listener.reportContextSensitivity(
-          recognizer, dfa, startIndex, stopIndex, prediction, configs);
+        recognizer,
+        dfa,
+        startIndex,
+        stopIndex,
+        prediction,
+        configs,
+      );
     }
   }
 }
