@@ -13,13 +13,17 @@ import 'token_source.dart';
 /// An [IntStream] whose symbols are [Token] instances.
 abstract class TokenStream extends IntStream {
   /// Get the [Token] instance associated with the value returned by
-  /// {@link #LA LA(k)}. This method has the same pre- and post-conditions as
-  /// {@link IntStream#LA}. In addition, when the preconditions of this method
+  /// [LA]. This method has the same pre- and post-conditions as
+  /// [IntStream.LA]. In addition, when the preconditions of this method
   /// are met, the return value is non-null and the value of
-  /// {@code LT(k).getType()==LA(k)}.
+  /// `LT(k).getType()==LA(k)`.
   ///
-  /// @see IntStream#LA
-  Token LT(int k);
+  /// TODO: in this doc it says that is non-null, but the implementation says
+  /// otherwise
+  ///
+  /// Se also:
+  /// ntStream.LA
+  Token? LT(int k);
 
   /// Gets the [Token] at the specified [index] in the stream. When
   /// the preconditions of this method are met, the return value is non-null.
@@ -114,7 +118,7 @@ abstract class TokenStream extends IntStream {
   ///
   /// @throws UnsupportedOperationException if this stream does not support
   /// this method for the specified tokens
-  String getTextRange(Token start, Token stop);
+  String getTextRange(Token? start, Token? stop);
 }
 
 /// This implementation of [TokenStream] loads tokens from a
@@ -159,11 +163,7 @@ class BufferedTokenStream implements TokenStream {
   /// <ul>
   bool fetchedEOF = false;
 
-  BufferedTokenStream(this._tokenSource) {
-    if (_tokenSource == null) {
-      throw ArgumentError.notNull('tokenSource');
-    }
-  }
+  BufferedTokenStream(this._tokenSource);
 
   @override
   int get index => p;
@@ -264,7 +264,8 @@ class BufferedTokenStream implements TokenStream {
   }
 
   /// Get all tokens from start..stop inclusively */
-  List<Token> getRange(int start, [int stop]) {
+  List<Token>? getRange(int start, [int? stop]) {
+    stop = stop ?? start;
     if (start < 0 || stop < 0) return null;
     lazyInit();
     final subset = <Token>[];
@@ -278,17 +279,17 @@ class BufferedTokenStream implements TokenStream {
   }
 
   @override
-  int LA(int i) {
-    return LT(i).type;
+  int? LA(int i) {
+    return LT(i)?.type;
   }
 
-  Token LB(int k) {
+  Token? LB(int k) {
     if ((p - k) < 0) return null;
     return tokens[p - k];
   }
 
   @override
-  Token LT(int k) {
+  Token? LT(int k) {
     lazyInit();
     if (k == 0) return null;
     if (k < 0) return LB(-k);
@@ -344,11 +345,16 @@ class BufferedTokenStream implements TokenStream {
   /// Given a start and stop index, return a List of all tokens in
   ///  the token type BitSet.  Return null if no tokens were found.  This
   ///  method looks at both on and off channel tokens.
-  List<Token> getTokens(
-      [int start, int stop, Set<int> types]) {
+  List<Token>? getTokens([
+    int? start,
+    int? stop,
+    Set<int>? types,
+  ]) {
     if (start == null && stop == null) {
       return tokens;
     }
+    start = start!;
+    stop = stop!;
     lazyInit();
     if (start < 0 || start >= tokens.length) {
       throw RangeError.index(start, tokens);
@@ -358,7 +364,7 @@ class BufferedTokenStream implements TokenStream {
     if (start > stop) return null;
 
     // list = tokens[start:stop]:{T t, t.getType() in types}
-    var filteredTokens = <Token>[];
+    List<Token>? filteredTokens = <Token>[];
     for (var i = start; i <= stop; i++) {
       final t = tokens[i];
       if (types == null || types.contains(t.type)) {
@@ -425,7 +431,7 @@ class BufferedTokenStream implements TokenStream {
   /// Collect all tokens on specified channel to the right of
   ///  the current token up until we see a token on DEFAULT_TOKEN_CHANNEL or
   ///  EOF. If channel is -1, find any non default channel token.
-  List<Token> getHiddenTokensToRight(int tokenIndex, [int channel = -1]) {
+  List<Token>? getHiddenTokensToRight(int tokenIndex, [int channel = -1]) {
     lazyInit();
     if (tokenIndex < 0 || tokenIndex >= tokens.length) {
       throw RangeError.index(tokenIndex, tokens);
@@ -443,7 +449,7 @@ class BufferedTokenStream implements TokenStream {
   /// Collect all tokens on specified channel to the left of
   ///  the current token up until we see a token on DEFAULT_TOKEN_CHANNEL.
   ///  If channel is -1, find any non default channel token.
-  List<Token> getHiddenTokensToLeft(int tokenIndex, [int channel = -1]) {
+  List<Token>? getHiddenTokensToLeft(int tokenIndex, [int channel = -1]) {
     lazyInit();
     if (tokenIndex < 0 || tokenIndex >= tokens.length) {
       throw RangeError.index(tokenIndex, tokens);
@@ -454,8 +460,10 @@ class BufferedTokenStream implements TokenStream {
       return null;
     }
 
-    final prevOnChannel =
-        previousTokenOnChannel(tokenIndex - 1, Lexer.DEFAULT_TOKEN_CHANNEL);
+    final prevOnChannel = previousTokenOnChannel(
+      tokenIndex - 1,
+      Lexer.DEFAULT_TOKEN_CHANNEL,
+    );
     if (prevOnChannel == tokenIndex - 1) return null;
     // if none onchannel to left, prevOnChannel=-1 then from=0
     final from = prevOnChannel + 1;
@@ -464,7 +472,7 @@ class BufferedTokenStream implements TokenStream {
     return filterForChannel(from, to, channel);
   }
 
-  List<Token> filterForChannel(int from, int to, int channel) {
+  List<Token>? filterForChannel(int from, int to, int channel) {
     final hidden = <Token>[];
     for (var i = from; i <= to; i++) {
       final t = tokens[i];
@@ -485,7 +493,7 @@ class BufferedTokenStream implements TokenStream {
   String get text => getText();
 
   @override
-  String getText([Interval interval]) {
+  String getText([Interval? interval]) {
     interval = interval ??
         Interval.of(0, size - 1); // Get the text of all tokens in this buffer.
     final start = interval.a;
@@ -509,7 +517,7 @@ class BufferedTokenStream implements TokenStream {
   }
 
   @override
-  String getTextRange(Token start, Token stop) {
+  String getTextRange(Token? start, Token? stop) {
     if (start != null && stop != null) {
       return getText(Interval.of(start.tokenIndex, stop.tokenIndex));
     }
@@ -578,7 +586,7 @@ class CommonTokenStream extends BufferedTokenStream {
   }
 
   @override
-  Token LB(int k) {
+  Token? LB(int k) {
     if (k == 0 || (p - k) < 0) return null;
 
     var i = p;
@@ -594,7 +602,7 @@ class CommonTokenStream extends BufferedTokenStream {
   }
 
   @override
-  Token LT(int k) {
+  Token? LT(int k) {
     //System.out.println("enter LT("+k+")");
     lazyInit();
     if (k == 0) return null;
