@@ -19,7 +19,12 @@ class Trees {
   /// Print out a whole tree in LISP form. {@link #getNodeText} is used on the
   ///  node payloads to get the text for the nodes.  Detect
   ///  parse trees and extract data appropriately.
-  static String toStringTree(Tree t, {Parser recog, List<String> ruleNames}) {
+  static String toStringTree(
+    Tree? t, {
+    Parser? recog,
+    List<String>? ruleNames,
+  }) {
+    if (t == null) return 'null';
     ruleNames ??= recog?.ruleNames;
     var s = escapeWhitespace(getNodeText(t, ruleNames: ruleNames), false);
     if (t.childCount == 0) return s;
@@ -36,7 +41,11 @@ class Trees {
     return buf.toString();
   }
 
-  static String getNodeText(Tree t, {Parser recog, List<String> ruleNames}) {
+  static String getNodeText(
+    Tree t, {
+    Parser? recog,
+    List<String>? ruleNames,
+  }) {
     ruleNames ??= recog?.ruleNames;
     if (ruleNames != null) {
       if (t is RuleContext) {
@@ -51,16 +60,14 @@ class Trees {
         return t.toString();
       } else if (t is TerminalNode) {
         final symbol = (t).symbol;
-        if (symbol != null) {
-          final s = symbol.text;
-          return s;
-        }
+        final s = symbol.text;
+        return s ?? '';
       }
     }
     // no recog for rule names
     Object payload = t.payload;
     if (payload is Token) {
-      return payload.text;
+      return payload.text ?? '';
     }
     return t.payload.toString();
   }
@@ -69,7 +76,7 @@ class Trees {
   static List<Tree> getChildren(Tree t) {
     final kids = <Tree>[];
     for (var i = 0; i < t.childCount; i++) {
-      kids.add(t.getChild(i));
+      kids.add(t.getChild(i)!);
     }
     return kids;
   }
@@ -80,11 +87,11 @@ class Trees {
   ///  @since 4.5.1
   static List<Tree> getAncestors(Tree t) {
     if (t.parent == null) return [];
+    var walker = t.parent;
     final ancestors = <Tree>[];
-    t = t.parent;
-    while (t != null) {
-      ancestors.insert(0, t); // insert at start
-      t = t.parent;
+    while (walker != null) {
+      ancestors.insert(0, walker); // insert at start
+      walker = walker.parent;
     }
     return ancestors;
   }
@@ -93,7 +100,7 @@ class Trees {
   ///  Use == not equals().
   ///
   ///  @since 4.5.1
-  static bool isAncestorOf(Tree t, Tree u) {
+  static bool isAncestorOf(Tree? t, Tree? u) {
     if (t == null || u == null || t.parent == null) return false;
     var p = u.parent;
     while (p != null) {
@@ -118,7 +125,11 @@ class Trees {
   }
 
   static void _findAllNodes(
-      ParseTree t, int index, bool findTokens, List<ParseTree> nodes) {
+    ParseTree t,
+    int index,
+    bool findTokens,
+    List<ParseTree> nodes,
+  ) {
     // check this node (the root) first
     if (findTokens && t is TerminalNode) {
       final tnode = t;
@@ -129,7 +140,7 @@ class Trees {
     }
     // check children
     for (var i = 0; i < t.childCount; i++) {
-      _findAllNodes(t.getChild(i), index, findTokens, nodes);
+      _findAllNodes(t.getChild(i)!, index, findTokens, nodes);
     }
   }
 
@@ -142,7 +153,7 @@ class Trees {
 
     final n = t.childCount;
     for (var i = 0; i < n; i++) {
-      nodes.addAll(getDescendants(t.getChild(i)));
+      nodes.addAll(getDescendants(t.getChild(i)!));
     }
     return nodes;
   }
@@ -156,25 +167,27 @@ class Trees {
   ///  inclusively using postorder traversal.  Recursive depth-first-search.
   ///
   ///  @since 4.5.1
-  static ParserRuleContext getRootOfSubtreeEnclosingRegion(
-      ParseTree t,
-      int startTokenIndex, // inclusive
-      int stopTokenIndex) // inclusive
-  {
+  static ParserRuleContext? getRootOfSubtreeEnclosingRegion(
+    ParseTree t,
+    int startTokenIndex, // inclusive
+    int stopTokenIndex, // inclusive
+  ) {
     final n = t.childCount;
     for (var i = 0; i < n; i++) {
-      final child = t.getChild(i);
+      final child = t.getChild(i)!;
       final r = getRootOfSubtreeEnclosingRegion(
-          child, startTokenIndex, stopTokenIndex);
+        child,
+        startTokenIndex,
+        stopTokenIndex,
+      );
       if (r != null) return r;
     }
     if (t is ParserRuleContext) {
-      final r = t;
-      if (startTokenIndex >=
-              r.start.tokenIndex && // is range fully contained in t?
-          (r.stop == null || stopTokenIndex <= r.stop.tokenIndex)) {
-        // note: r.getStop()==null likely implies that we bailed out of parser and there's nothing to the right
-        return r;
+      // is range fully contained in t?
+      final start = t.start != null && startTokenIndex >= t.start!.tokenIndex;
+      final end = t.stop == null || stopTokenIndex <= t.stop!.tokenIndex;
+      if (start && end) {
+        return t;
       }
     }
     return null;
@@ -187,18 +200,22 @@ class Trees {
   ///  WARNING: destructive to t.
   ///
   ///  @since 4.5.1
-  static void stripChildrenOutOfRange(ParserRuleContext t,
-      ParserRuleContext root, int startIndex, int stopIndex) {
+  static void stripChildrenOutOfRange(
+    ParserRuleContext? t,
+    ParserRuleContext root,
+    int startIndex,
+    int stopIndex,
+  ) {
     if (t == null) return;
     for (var i = 0; i < t.childCount; i++) {
-      final child = t.getChild(i);
+      final child = t.getChild(i)!;
       final range = child.sourceInterval;
       if (child is ParserRuleContext &&
           (range.b < startIndex || range.a > stopIndex)) {
         if (isAncestorOf(child, root)) {
           // replace only if subtree doesn't have displayed root
           final abbrev = CommonToken(Token.INVALID_TYPE, text: '...');
-          t.children[i] = TerminalNodeImpl(abbrev);
+          t.children![i] = TerminalNodeImpl(abbrev);
         }
       }
     }
@@ -207,7 +224,7 @@ class Trees {
   /// Return first node satisfying the pred
   ///
   ///  @since 4.5.1
-  static Tree findNodeSuchThat(Tree t, Predicate<Tree> pred) {
+  static Tree? findNodeSuchThat(Tree? t, Predicate<Tree> pred) {
     if (pred.test(t)) return t;
 
     if (t == null) return null;
@@ -222,5 +239,5 @@ class Trees {
 }
 
 abstract class Predicate<T> {
-  bool test(T t);
+  bool test(T? t);
 }
